Rebuilt authentication logic
Had to restructure some of the authentication logic to allow wildcard restrictions on roles for endpoints. It is now for example possible setup scopes such as 'roles:company_admin:*' basically indicating that user has to be a company admin to perform said operation but it doesn't matter for what company. The exception to this is the 'self' claim, where wildcard is not allowed. Self will always mean self only. Nothing else would make sense.