"Updated Django version due to vulterability warning from github."
"Added try catch for a couple of forms definitions for migrate command to work when being applied to a new db. Not sure why this should be needed but I could find no other solution. If there isn't one, then I have to say this is a severe oversight by the django developers..."